+++ client/package.json
@@ -8,7 +8,7 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@girder/components": "git+https://github.com/girder/girder_web_components.git",
+    "@girder/components": "^2.2.1",
     "core-js": "^2.6.5",
     "d3": "^5.12.0",
     "geojs": "^0.19.5",
