+++ client/src/components/annotators/VideoAnnotator.vue
@@ -1,57 +1,25 @@
 <script>
-import { throttle } from "lodash";
-
-import Vue from "vue";
-import geo from "geojs";
+import annotator from "./annotator";
 
 export default {
   name: "VideoAnnotator",
+
   props: {
     videoUrl: {
       type: String,
       required: true
-    },
-    frameRate: {
-      type: Number,
-      required: true
     }
   },
-  provide() {
-    return {
-      annotator: this.provided
-    };
-  },
-  data() {
-    this.provided = new Vue({
-      computed: {
-        viewer: () => this.viewer,
-        playing: () => this.playing,
-        frame: () => this.frame,
-        maxFrame: () => this.maxFrame,
-        syncedFrame: () => this.syncedFrame
-      }
-    });
-    return {
-      ready: false,
-      playing: false,
-      frame: 0,
-      maxFrame: 0,
-      syncedFrame: 0
-    };
-  },
+
+  mixins: [annotator],
+
   computed: {
     isLastFrame() {
       return this.playing;
     }
   },
+
   created() {
-    this.provided.$on("play", this.play);
-    this.provided.$on("prev-frame", this.prevFrame);
-    this.provided.$on("next-frame", this.nextFrame);
-    this.provided.$on("pause", this.pause);
-    this.provided.$on("seek", this.seek);
-    this.emitFrame();
-    this.emitFrame = throttle(this.emitFrame, 200);
     var video = document.createElement("video");
     this.video = video;
     video.preload = "auto";
@@ -65,38 +33,10 @@ export default {
     };
     video.addEventListener("pause", this.videoPaused);
   },
+
   methods: {
     init() {
-      var params = geo.util.pixelCoordinateParams(
-        this.$refs.container,
-        this.width,
-        this.height,
-        this.width,
-        this.height
-      );
-      this.viewer = geo.map(params.map);
-      this.viewer.zoomRange({
-        min: this.viewer.zoomRange().origMin,
-        max: this.viewer.zoomRange().max + 3
-      });
-      var interactorOpts = this.viewer.interactor().options();
-      interactorOpts.keyboard.focusHighlight = false;
-      interactorOpts.keyboard.actions = {};
-      interactorOpts.actions = [
-        interactorOpts.actions[0],
-        interactorOpts.actions[2],
-        interactorOpts.actions[6],
-        interactorOpts.actions[7],
-        interactorOpts.actions[8]
-      ];
-      interactorOpts.zoomAnimation = {
-        enabled: false
-      };
-      interactorOpts.momentum = {
-        enabled: false
-      };
-      interactorOpts.wheelScaleY = 0.2;
-      this.viewer.interactor().options(interactorOpts);
+      this.baseInit();
       this.quadFeatureLayer = this.viewer.createLayer("feature", {
         features: ["quad.video"]
       });
@@ -112,15 +52,17 @@ export default {
         .draw();
       this.ready = true;
     },
+
     async play() {
       try {
         await this.video.play();
         this.playing = true;
         this.syncWithVideo();
       } catch (ex) {
-        console.error(ex);
+        console.log(ex);
       }
     },
+
     async seek(frame) {
       this.video.currentTime = frame / this.frameRate;
       this.frame = Math.round(this.video.currentTime * this.frameRate);
@@ -128,32 +70,25 @@ export default {
       this.video.removeEventListener("seeked", this.pendingUpdate);
       this.video.addEventListener("seeked", this.pendingUpdate);
     },
+
     pendingUpdate() {
       this.syncedFrame = Math.round(this.video.currentTime * this.frameRate);
     },
-    prevFrame() {
-      var targetFrame = this.frame - 1;
-      if (targetFrame >= 0) {
-        this.seek(targetFrame);
-      }
-    },
-    nextFrame() {
-      var targetFrame = this.frame + 1;
-      if (targetFrame <= this.maxFrame) {
-        this.seek(targetFrame);
-      }
-    },
+
     pause() {
       this.video.pause();
       this.playing = false;
     },
+
     videoPaused() {
       if (this.video.currentTime === this.video.duration) {
+        // console.log("video ended");
         this.frame = 0;
         this.syncedFrame = 0;
         this.pause();
       }
     },
+
     onResize() {
       if (!this.viewer) {
         return;
@@ -164,17 +99,14 @@ export default {
         this.viewer.size(size);
       }
     },
+
     syncWithVideo() {
       if (this.playing) {
         this.frame = Math.round(this.video.currentTime * this.frameRate);
         this.syncedFrame = this.frame;
         this.viewer.scheduleAnimationFrame(this.syncWithVideo);
       }
-    },
-    emitFrame() {
-      this.$emit("frame-update", this.frame);
-    },
-    rendered() {}
+    }
   }
 };
 </script>
@@ -188,23 +120,5 @@ export default {
 </template>
 
 <style lang="scss" scoped>
-.video-annotator {
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  z-index: 0;
-
-  display: flex;
-  flex-direction: column;
-
-  .playback-container {
-    flex: 1;
-
-    &.geojs-map:focus {
-      outline: none;
-    }
-  }
-}
+@import "./annotator.scss";
 </style>
