+++ client/src/components/annotators/annotator.scss
@@ -1,0 +1,19 @@
+.video-annotator {
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: 0;
+  bottom: 0;
+  z-index: 0;
+
+  display: flex;
+  flex-direction: column;
+
+  .playback-container {
+    flex: 1;
+
+    &.geojs-map:focus {
+      outline: none;
+    }
+  }
+}
