+++ client/src/components/layers/AnnotationLayer.vue
