+++ client/src/components/layers/EditAnnotationLayer.vue
