+++ client/src/components/layers/PointMarkingLayer.vue
